New maven-built project, using TooTallNate's websocket server
If we used Java’s EE websockets we’d need Tomcat or Glassfish or
something else entirely too complex and annoying.

https://github.com/TooTallNate/Java-WebSocket/

This is a standalone websocket server, which means all we need is Maven
and we can run it by itself without all this other crap.

I tried Spark but there were a couple errors and this is fine as is.
It’s not going to be too central to the app.